APPGATE:SOLID for the storage manager.
